Move user service folder
