+++ .env.sample
@@ -10,7 +10,7 @@ QUESTION_SERVICE_PORT=8081
 
 # User Service
 USER_SERVICE_CLOUD_URI=mongodb+srv://admin:<db_password>@cluster0.uo0vu.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
-USER_SERVICE_LOACL_URI=mongodb://127.0.0.1:27017/peerprepUserServiceDB
+USER_SERVICE_LOCAL_URI=mongodb://127.0.0.1:27017/peerprepUserServiceDB
 USER_SERVICE_PORT=3001
 
 # Will use cloud MongoDB Atlas database
