--- services/user-service/.env.sample
@@ -1,12 +1,0 @@
-# User Service
-USER_SERVICE_CLOUD_URI=<cloud_uri>
-USER_SERVICE_LOACL_URI=mongodb://127.0.0.1:27017/peerprepUserServiceDB
-USER_SERVICE_PORT=3001
-
-# Will use cloud MongoDB Atlas database
-ENV=PROD
-
-# Secret for creating JWT signature
-JWT_SECRET=you-can-replace-this-with-your-own-secret
-
-NODE_ENV=development
