--- services/user-service/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "user-service",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./src/index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "nodemon ./src/server.js",
-    "start": "node ./src/server.js",
-    "test": "mocha --require ts-node/register tests/**/*.spec.ts --exit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/chai": "^5.0.0",
-    "@types/mocha": "^10.0.8",
-    "@types/node": "^22.6.1",
-    "chai": "^5.1.1",
-    "chai-http": "^5.0.0",
-    "mocha": "^10.7.3",
-    "mongodb-memory-server": "^10.0.1",
-    "nodemon": "^3.1.4",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.6.2"
-  },
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.5.4"
-  }
-}
