+++ services/user/MongoDBSetup.md
@@ -1,0 +1,60 @@
+# Setting up MongoDB Instance for User Service
+
+1. Visit the MongoDB Atlas Site [https://www.mongodb.com/atlas](https://www.mongodb.com/atlas) and click on "Try Free"
+
+2. Sign Up/Sign In with your preferred method.
+
+3. You will be greeted with welcome screens. Feel free to skip them till you reach the Dashboard page.
+
+4. Create a Database Deployment by clicking on the green `+ Create` Button:
+
+![alt text](./GuideAssets/Creation.png)
+
+5. Make selections as followings:
+
+- Select Shared Cluster
+- Select `aws` as Provider
+
+![alt text](./GuideAssets/Selection1.png)
+
+- Select `Singapore` for Region
+
+![alt text](./GuideAssets/Selection2.png)
+
+- Select `M0 Sandbox` Cluster (Free Forever - No Card Required)
+
+> Ensure to select M0 Sandbox, else you may be prompted to enter card details and may be charged!
+
+![alt text](./GuideAssets/Selection3.png)
+
+- Leave `Additional Settings` as it is
+
+- Provide a suitable name to the Cluster
+
+![alt text](./GuideAssets/Selection4.png)
+
+6. You will be prompted to set up Security for the database by providing `Username and Password`. Select that option and enter `Username` and `Password`. Please keep this safe as it will be used in User Service later on.
+
+![alt text](./GuideAssets/Security.png)
+
+7. Next, click on `Add my Current IP Address`. This will whiteliste your IP address and allow you to connect to the MongoDB Database.
+
+![alt text](./GuideAssets/Network.png)
+
+8. Click `Finish and Close` and the MongoDB Instance should be up and running.
+
+## Whitelisting All IP's
+
+1. Select `Network Access` from the left side pane on Dashboard.
+
+![alt text](./GuideAssets/SidePane.png)
+
+2. Click on the `Add IP Address` Button
+
+![alt text](./GuideAssets/AddIPAddress.png)
+
+3. Select the `ALLOW ACCESS FROM ANYWHERE` Button and Click `Confirm`
+
+![alt text](./GuideAssets/IPWhitelisting.png)
+
+Now, any IP Address can access this Database.
