+++ services/user/index.js
@@ -1,0 +1,59 @@
+import express from "express";
+import cors from "cors";
+
+import userRoutes from "./src/routes/user-routes.js";
+import authRoutes from "./src/routes/auth-routes.js";
+
+const app = express();
+
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
+app.use(cors()); // config cors so that front-end can use
+app.options("*", cors());
+
+// To handle CORS Errors
+app.use((req, res, next) => {
+  res.header("Access-Control-Allow-Origin", "*"); // "*" -> Allow all links to access
+
+  res.header(
+    "Access-Control-Allow-Headers",
+    "Origin, X-Requested-With, Content-Type, Accept, Authorization",
+  );
+
+  // Browsers usually send this before PUT or POST Requests
+  if (req.method === "OPTIONS") {
+    res.header("Access-Control-Allow-Methods", "GET, POST, DELETE, PUT, PATCH");
+    return res.status(200).json({});
+  }
+
+  // Continue Route Processing
+  next();
+});
+
+app.use("/users", userRoutes);
+app.use("/auth", authRoutes);
+
+app.get("/", (req, res, next) => {
+  console.log("Sending Greetings!");
+  res.json({
+    message: "Hello World from user-service",
+  });
+});
+
+// Handle When No Route Match Is Found
+app.use((req, res, next) => {
+  const error = new Error("Route Not Found");
+  error.status = 404;
+  next(error);
+});
+
+app.use((error, req, res, next) => {
+  res.status(error.status || 500);
+  res.json({
+    error: {
+      message: error.message,
+    },
+  });
+});
+
+export default app;
